+++ lab3/src/main/java/org/lab3/model/objects/SlashBladeObject.java
@@ -1,8 +1,8 @@
 package org.lab3.model.objects;
 
 public interface SlashBladeObject {
-    void setInGamePosition(int x, int y);
-    void changeInGamePosition(int x, int y);
-    int getInGamePosX();
-    int getInGamePosY();
+    void setInGamePosition(double x, double y);
+    void changeInGamePosition(double x, double y);
+    double getInGamePosX();
+    double getInGamePosY();
 }
